create get for pet (status, tag, id). create test for class URLBulder

+++ src/main/java/models/Order.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class Order {
-    private int id;
-    private int petId;
-    private int quantity;
+    private long id;
+    private long petId;
+    private long quantity;
     private String shipDate;
     private Status status;
     private boolean complete;
